[Refactor] Move ts folder
