+++ ts/commonUI/karelStyleModal.ts
