+++ ts/desktop/console.ts
