+++ ts/desktop/controlBar.ts
