+++ ts/desktop/focusBar.ts
