+++ ts/editor/themes/classicHighlight.ts
