+++ ts/editor/themes/darkClassicHighlight.ts
