+++ ts/editor/themes/editorTheme.ts
