+++ ts/index.ts
@@ -3,8 +3,8 @@ import { SelectLine, setLanguage, SetText } from "./editor/editor";
 import { DesktopController } from "./desktop/desktop-ui";
 import { GetPhoneUIHelper } from "./phone-ui";
 import { HookUpCommonUI } from "./common-ui";
-import { World } from "../../js/karel";
-import { karel } from "../../js";
+import { World } from "../js/karel";
+import { karel } from "../js";
 import { KarelController } from "./KarelController";
 import { responsiveHack } from "./responsive-load";
 import { InitSettings, StartSettings } from "./settings";
