+++ ts/volatileMemo.ts
