+++ ts/worldViewController/waffle.ts
