--- webapp/test.html
@@ -1,33 +1,0 @@
-<!DOCTYPE HTML>
- <html>
- <head>
- <style type="text/css">
-       #div1 {width:350px;height:70px;padding:10px;border:1px solid #aaaaaa;}
- </style>
- </head>
- <body>
-
- <p>Drag the W3Schools image into the rectangle:</p>
-
- <div id="div1"></div>
- <br>
- <img id="drag" src="img_logo.gif" width="336" height="69">
-
- <script type="text/javascript">
-  var el = document.getElementById('drag'); 
-
-  el.addEventListener("touchstart", handleStart, false);
-  el.addEventListener("touchend", handleEnd, false);
-  el.addEventListener("touchcancel", handleCancel, false);
-  el.addEventListener("touchleave", handleEnd, false);
-  el.addEventListener("touchmove", handleMove, false);
-
-  function handleStart(event) {
-      // Handle the start of the touch
-  }
-
-  // ^ Do the same for the rest of the events
-
-</script>
-</body>
-</html>
